[UPDATE] - create a folder widget and move all widgets on it,  create a utils function for starts evaluation
