+++ mobile_app/lib/widgets/assistant.dart
